More tests. testOutputWithPreAndScriptTags is now correct, I think. The overarching goal of StringBean is to return the text, and only the text, that a user would see in a browser. When I open this html in Netscape I see nothing:
<body><pre>
<script language=\"javascript\">
if(navigator.appName.indexOf(\"Netscape\") != -1)
 document.write ('xxx');
else
 document.write ('yyy');
</script>
</pre></body>

That's what StringBean says too.

Although this points out a flaw in the specification (such as it is, see the StringBean javadoc):

 * Text within <SCRIPT></SCRIPT> tags is removed.
 * The text within <PRE></PRE> tags is not altered.

So what happens when there's both?
The 'not altered' means, not collapsed, not that the HTML is returned literally.
To test this out, see how Netscape handles:

<body><pre>
<b>Hello
World</b>
</pre></body>

For me it shows:

Hello
World

which means it consumes (and honours) the bold tags even though it is within <pre></pre>
So the StringBean <pre> handling is correct, or so it seems.

I guess that most questions can be answered by submitting it to a browser.