Fixed bug #757337 Standalone attributes should remain standalone
Changing to a HashMap was not adopted for two reasons. The first is that it would break Java 1.1.x
compatibility (which I believe we still maintain, although I haven't checked lately), and using a HashMap
would return the attributes in a different order, leading to a *lot* of broken tests that rely on toHTML()
returning a specific string with attributes in a certain order (if I'm going to fix a bunch of tests I would
like the attributes to come out in the order they were originally, so this will need an AttributeStorage class
when the attribute parser is rewritten).

The solution adopted is to subclass HashTable and store two special values, which are new constants in
the Tag class.  For naive programs, and for backward compatibility, this SpecialHashtable class translates
these constants into null and "" respectively, so the old behavior is the same.

However, savvy programs, and the toHTML() method in the Tag class can call getRaw() on this hashtable
to get at these special constants and behave appropriately. For this specific bug, toHTML returns the
stand-alone attribute as is, but the missing value case is also handled.