Add style checking target to ant build script:     ant checkstyle It uses a jar from http://checkstyle.sourceforge.net which is dropped in the lib directory. The rules are in the file htmlparser_checks.xml in the src directory.
Added lexerapplications package with Tabby as the first app. It performs whitespace manipulation
on source files to follow the style rules. This reduced the number of style violations to roughly 14,000.

There are a few issues with the style checker that need to be resolved before it should be taken too seriously.
For example:
It thinks all method arguments should be final, even if they are modified by the code (which the compiler frowns on).
It complains about long lines, even when there is no possibility of wrapping the line, i.e. a URL in a comment
that's more than 80 characters long.
It considers all naked integers as 'magic numbers', even when they are obvious, i.e. the 4 corners of a box.
It complains about whitespace following braces, even in array initializers, i.e. X[][] = { {a, b} { } }

But it points out some really interesting things, even if you don't agree with the style guidelines,
so it's worth a look.